Moved components to folders.
