+++ src/views/DebitView.vue
@@ -1,7 +1,7 @@
 <script setup>
-import CardsTitle from '../components/CardsTitle.vue';
-import CardDescription from '../components/CardDescription.vue';
-import OrderFormDebit from '../components/OrderFormDebit.vue';
+import CardsTitle from '../components/cards/CardsTitle.vue';
+import CardDescription from '../components/cards/CardDescription.vue';
+import OrderFormDebit from '../components/cards/OrderFormDebit.vue';
 // import CardsButtons from '../components/CardsButtons.vue'
 import GuidesMain from '../components/GuidesMain.vue';
 </script>
@@ -10,6 +10,6 @@ import GuidesMain from '../components/GuidesMain.vue';
     <CardsTitle cardType="Debit" cardDescription="Using debit card is the best way to pay for everything you want and after all - receive a cashback"></CardsTitle>
     <!-- <CardsButtons cardName="4U Debit" cardType="Debit"></CardsButtons> -->
     <CardDescription cardName="Galaxy" description="Debit card that you can use everywhere you want and take all advantages of it without any problems" imgLink="/src/assets/images/cards/CardDebet.svg" ></CardDescription>
-    <GuidesMain pageTitle="Conditions" fCardTitle="Cashback" fCardDescription="Up to 15% of cashback from each payment" fImgLink="/src/assets/images/icons/cards-description/cashback.svg" sCardTitle="Deposit" sCardDescription="Up to 5% annual for saving on your card" sImgLink="/src/assets/images/icons/cards-description/deposit.svg" tCardTitle="Commission" tCardDescription="No taxes and fees for using your cards" tImgLink="/src/assets/images/icons/cards-description/commission.svg"></GuidesMain>
+    <GuidesMain pageTitle="Conditions" fCardTitle="Cashback" fCardDescription="Up to 30% of cashback from each payment" fImgLink="/src/assets/images/icons/cards-description/cashback.svg" sCardTitle="Deposit" sCardDescription="Up to 2% annual for saving on your card" sImgLink="/src/assets/images/icons/cards-description/deposit.svg" tCardTitle="Commission" tCardDescription="No taxes and fees for using your cards" tImgLink="/src/assets/images/icons/cards-description/commission.svg"></GuidesMain>
     <OrderFormDebit></OrderFormDebit>
 </template>
